Remove check-cluster-readiness job
It was deploying an app to make sure the cluster was ready, but we
already deploy couple of things on cluster in prepare-cluster job, so it
was redundant. We started hardening the cluster security by making every
service account to require explicit PSP use access and maintaining this
job was not worth the effort.

[#167322148]

Signed-off-by: Akshay Mankar <amankar@pivotal.io>