Change custom role name to use camel case
One of the roles (eatsstable) fails to create due to it being marked for
deletion. All the other roles seem to be alright. We can also not find
this role even with the --show-deleted option in gcloud. For now we'll
just rename them using camel case and see if it fails next time the
clusters are recreated.

Co-authored-by: Mario Nitchev <marionitchev@gmail.com>